feat: implement trade exit engine and strengthen paper trading validation

+++ signal-engine-service/src/main/java/com/ram/trading/signal/engine/exit/ExitDecision.java
@@ -16,7 +16,7 @@ public class ExitDecision {
 
     private ExitReason reason;
 
-    private BigDecimal exitPrice;
+    private Double exitPrice;
 
     private ExitType exitType;
